requirements.txt
# UniCardio runtime dependencies.
#
# 注意 torch 不在本文件里：CUDA 版本和平台强相关，请先按下面的命令安装，再装本文件。
# 项目源码不用 torchvision / torchaudio，省略。
#
#   # 服务器 (CUDA 12.8, H800) —— 当前生产环境
#   pip install torch==2.11.0 --index-url https://download.pytorch.org/whl/cu128
#
#   # macOS CPU 调试（同版本，走默认 PyPI）
#   pip install torch==2.11.0
#
# 然后：
#   pip install -r requirements.txt

hydra-core>=1.3,<1.4
omegaconf>=2.3,<2.4
numpy>=1.24,<2.0
pandas>=1.5,<3.0
scikit-learn>=1.2,<2.0
tqdm>=4.65
swanlab>=0.7,<1.0

# 自带 warmup 的 cosine 调度器：trainer.py::_build_scheduler
cosine_annealing_warmup>=2.0,<3

# script/plot_first_sample.py 用到；不画图可不装
matplotlib>=3.7

# evaluate.py / trainer.py 的 SBP/DBP 提取（src/utils/bp_metrics.py）
pyvital>=0.5

# script/convert_h5_to_npy_csv.py 用到；不跑那个脚本可以不装
h5py>=3.8

# dev
pytest>=7.0
